Refactor analysis status.
